feat: implements the new DeleteUnixFiles DTO

+++ src/domain/useCase/deleteUnixFiles.go
@@ -6,12 +6,13 @@ import (
 
 	"github.com/speedianet/os/src/domain/dto"
 	"github.com/speedianet/os/src/domain/repository"
+	"github.com/speedianet/os/src/domain/valueObject"
 )
 
 func DeleteUnixFiles(
 	filesQueryRepo repository.FilesQueryRepo,
 	filesCmdRepo repository.FilesCmdRepo,
-	deleteUnixFiles dto.DeleteUnixFile,
+	deleteUnixFiles dto.DeleteUnixFiles,
 ) {
 	for fileToDeleteIndex, fileToDelete := range deleteUnixFiles.SourcePaths {
 		isRootPath := fileToDelete.String() == "/"
@@ -31,5 +32,32 @@ func DeleteUnixFiles(
 		deleteUnixFiles.SourcePaths = filesToDeleteWithoutNotAllowedPath
 	}
 
-	filesCmdRepo.Delete(deleteUnixFiles)
+	if deleteUnixFiles.PermanentDelete {
+		filesCmdRepo.Delete(deleteUnixFiles)
+		return
+	}
+
+	for _, fileToMoveToTrash := range deleteUnixFiles.SourcePaths {
+		trashPath, _ := valueObject.NewUnixFilePath("/.trash")
+		_, err := filesQueryRepo.GetOne(trashPath)
+		if err != nil {
+			log.Print("TrashNotFound")
+		}
+
+		updateUnixFile := dto.NewUpdateUnixFile(
+			fileToMoveToTrash,
+			&trashPath,
+			nil,
+			nil,
+		)
+
+		err = filesCmdRepo.Move(updateUnixFile)
+		if err != nil {
+			log.Printf(
+				"MoveUnixFileToTrashError (%s): %s",
+				fileToMoveToTrash.String(),
+				err.Error(),
+			)
+		}
+	}
 }
